Torneos finalizado 100funcional

+++ tournaments/models.py
@@ -1,5 +1,7 @@
 from django.db import models
 from .choices import estado
+from firebase_admin import firestore
+
 # Create your models here.
 
 class Game(models.Model):
@@ -25,15 +27,17 @@ class Game(models.Model):
 
 class Tournament(models.Model):
     id = models.AutoField(primary_key=True)
+    uidRegister = models.CharField( null=True, blank=True, max_length=100, verbose_name='UID Firestore')
     game = models.ForeignKey(Game, null=True, blank=True, on_delete=models.CASCADE, verbose_name= 'Juego')
     name = models.CharField(max_length=100, verbose_name= 'Nombre')
     status = models.CharField(max_length=7, choices=estado, default='C', verbose_name= 'Estado')
     date = models.DateField( verbose_name= 'Fecha')
     time = models.TimeField( verbose_name= 'Hora')
-    max_participants = models.IntegerField( verbose_name= 'Participantes')
+    players = models.IntegerField( null=True, blank=True, verbose_name= 'N° participantes')
     points = models.IntegerField( verbose_name= 'Puntaje')
     rules = models.TextField(max_length=3000, verbose_name= 'Reglamento')
     url = models.CharField(max_length=1000,  verbose_name= 'Link')
+    imageURL = models.CharField(null=True, blank=True, max_length=1000,  verbose_name= 'Link Imagen Firestore')
     image = models.ImageField(upload_to='tournaments/images/', verbose_name= 'Imagen', null=True)
 
     def __str__(self):
@@ -42,9 +46,26 @@ class Tournament(models.Model):
         fila = 'Titulo: ' + self.name + ' - ' + 'Fecha: ' + fecha + ' ' + 'Hora: ' + hora
         return fila
     
+    def save(self, *args, **kwargs):
+        if self.pk:
+            original = Tournament.objects.get(pk=self.pk)
+            self.uidRegister = original.uidRegister
+            self.imageURL = original.imageURL
+        super().save(*args, **kwargs)
+    
     def delete(self, using=None, keep_parents=False):
-        self.image.storage.delete(self.image.name)
-        super().delete()
+        # Elimina la imagen asociada al torneo si existe
+        if self.image:
+            self.image.storage.delete(self.image.name)
+        
+        # Obtiene la referencia al documento del torneo en Firestore usando el UID almacenado
+        tournament_ref = firestore.client().collection('tournaments').document(self.uidRegister)
+        
+        # Elimina el documento en Firestore
+        tournament_ref.delete()
+        
+        # Llama al método delete de la superclase para realizar la eliminación en Django
+        super().delete(using, keep_parents)
     class Meta:
         verbose_name = 'tournament'
         verbose_name_plural = 'tournaments'
@@ -62,6 +83,8 @@ class Player(models.Model):
     def __str__(self):
         fila = 'Nombre: ' + self.username + ' - ' + 'Puntaje: ' + str(self.points)
         return fila
+    
+
 
     def delete(self, using=None, keep_parents=False):
         self.image.storage.delete(self.image.name)
